Let CMake generate the master header file dynamically
